Update to docs with references to new deployment process
This addresses and closes #421

The chime-live-cluster.md file which contained release instructions has been shorted to just to explain how to access the cluster.

The prepar-new-chime-release.md is the starting point

The release document has been updated, and will show how to do the release

Because the process is two parts, the documentation represents how it is currently implemented.