Add st_app.py to the Docker file.
The getting started process is broken because the dev is expected to run the script/server script and have everything work.
Upon looking at the Dockerfile the app is in a constant restart cycle and therefore can't start.

Looking at the logs the st_app.py file is not being copied to the workdir in the container and therefore the container fails to start.